requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.8.1
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.3.1
cryptography==46.0.6
distro==1.9.0
dj-database-url==3.1.0
Django==4.2.13
django-email-users==2.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
dnspython==2.6.1
dotenv==0.9.9
filelock==3.25.2
Flask==3.0.3
Flask-Mail==0.10.0
Flask-SocketIO==5.3.7
Flask-Sockets==0.2.1
fsspec==2026.3.0
generativeai==0.0.1
gevent==24.10.2
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
greenlet==3.1.1
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.13.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.2.6
openai==2.29.0
packaging==26.0
pandas==2.2.3
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
rich==14.3.3
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.0
